+++ astro/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @bs-core/astro
 
+## 1.4.7
+
+### Patch Changes
+
+- Updated \_app.render to use the new method signature
+
+## 1.4.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @bs-core/shell@1.16.4
+
+## 1.4.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @bs-core/shell@1.16.3
+
+## 1.4.4
+
+### Patch Changes
+
+- Added immutableRegex to the Astro adapter config
+
 ## 1.4.3
 
 ### Patch Changes