+++ plugins/ldap/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @bs-plugins/ldap
 
+## 1.2.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @bs-core/shell@1.16.4
+
+## 1.2.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @bs-core/shell@1.16.3
+
 ## 1.2.3
 
 ### Patch Changes