+++ shell/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @bs-core/shell
 
+## 1.16.4
+
+### Patch Changes
+
+- Fixed issue when the immutable regex on the static file server
+
+## 1.16.3
+
+### Patch Changes
+
+- Added check to see if this.\_immutableRegex is undefined
+
 ## 1.16.2
 
 ### Patch Changes